requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Deprecated==1.2.12
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
PyGithub==1.54.1
PyJWT==1.7.1
requests==2.25.1
soupsieve==2.2.1
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
